+++ lib/src/io/plist_gen.dart
@@ -1,10 +1,17 @@
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:dcli/dcli.dart';
 
 import '../../flutter_translation_sheet.dart';
 
+/// Keeps the iOS Info.plist locales in sync with your trconfig.yaml locales.
+/// Only runs on MacOs, as some Linux distributions have a different `plutil`
+/// version that throws with the arguments.
 void addLocalesInPlist() {
+  if (!Platform.isMacOS) {
+    return;
+  }
   // replaces locales in the InfoPlist if available (only on Macos)
   // <key>CFBundleLocalizations</key>
   // <array>